requirements.txt
pillow
streamlit 